+++ documentation/ChangeLog.md
@@ -5,6 +5,8 @@
  * Fixed: can't save preference pages with empty fields.
  * Added support for source formatting using rustfmt (`Ctrl+Shift+F`).
   * Added `Format Crate (cargo fmt)` command to Project context menu.
+ * Fixed: Preferences "Download..." button not working in Windows if HOME env-var not set (now uses `%UserProfile%` instead).
+  * Also, "Download..." button now tries to download to the existing location in the preferences field, if it's a valid one.
  
 
 ### 0.5.0